Moving fhirpath.yml and utils file
