+++ containers/ecr-viewer/src/app/api/utils.ts
@@ -4,10 +4,7 @@ import yaml from "js-yaml";
 import { PathMappings } from "@/app/utils";
 
 export function loadYamlConfig(): PathMappings {
-  const filePath = path.join(
-    process.cwd(),
-    "src/app/api/fhir-data/fhirPath.yml",
-  );
+  const filePath = path.join(process.cwd(), "src/app/api/fhirPath.yml");
   const fileContents = fs.readFileSync(filePath, "utf8");
   return <PathMappings>yaml.load(fileContents);
 }
