+++ containers/ecr-viewer/src/app/view-data/components/tests/EcrSummary.test.tsx
@@ -9,7 +9,7 @@ describe("EcrSummary", () => {
   let container: HTMLElement;
   beforeAll(() => {
     const fhirPathFile = fs
-      .readFileSync("./src/app/api/fhir-data/fhirPath.yml", "utf8")
+      .readFileSync("./src/app/api/fhirPath.yml", "utf8")
       .toString();
     const fhirPathMappings = YAML.parse(fhirPathFile);
     const fhirBundle: Bundle = JSON.parse(
